requirements.txt
pypdf>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
streamlit>=1.20.0
pandas>=1.5.0
# Optional: for free local embeddings (Hugging Face sentence-transformers)
sentence-transformers>=2.2.0
